Add files via upload
This system allows employees to request equipment through a digital process, eliminating the need to do it on paper or in person. There are two types of users in the system: admins and regular users. The 'admin page' allows admins to add or remove users/employees, and to accept or decline user requests. The 'user page' enables users to submit requests for the equipment they need.